utils/requirements.txt
tensorflow-macos
tensorflow-metal
Pillow